Fixed amber patch
Before this fix, amber was working correctly with multiple
replicas but not working correctly with a single replica.

Now both situations should be ok.

Notice that I used a hack to detect if replicas are present
or not in sander. Probably this should be cleaned up in
some way

(cherry picked from commit 0b7dc44635708e821e14ed6f1156317807e6db7d)