Fixed bug for lammps running on 1 node
Lammps always shuffles the atom indices even when it is not running
on multiple processors.  As a context it is always important to use
the gatindex when collecting atom positions.  Previously this was
not done but now a check is added to ensure this is done if it is
needed.