Fixing regtests to use NOPBC
Now COM CENTER and GYRATION need NOPBC flag to work in the old way
(that is, without applying PBCs). I went through the regtests and
added NOPBC whenever necessary to make then work.
In real cases, this eventuality should
be very uncommon, since most of the people are using COM CENTER and
GYRATION on a whole molecule (likely made whole with a previous
WHOLEMOLECULES command).