Changed FitToTemplate::apply()
It does not shift anymore the atoms to their original positions.

Notice that this makes its behavior consistent with WHOLEMOLECULES
where atoms are not restored in the backward loop.

This implies that (consistently) one cannot rely that shared atoms
(those accessed with modifyAtoms()) are in their original position
in the ::apply() methods.