PBC slightly faster on orthorombic cells
Taken from the fccubic branch of @ceriottm

The "while" implementation for orthorombic cells is slighlty faster,
so I decided to enable it by default.

The "while" implementation for Tools::pbc is either slower
or equivalent, so I did not include it.

I think the reason is that in the first case one can avoid a couple
of multiplications, whereas in the second this is more difficult
since images are implemented in scaled coordinates and require
anyway two matrix*vector multiplications.