Implemented possibility to pass restart flag from MD
Added a new cmd "setRestart" that allow setting restart
from MD engine.

This setting can be overridden by PLUMED using:
RESTART
or
RESTART OFF
to respectively switch on or off restart from PLUMED input.

Notice that this would break compatibility in the sense that
it would turn on restart automatically without the user
specifying it.

I increased the APIversion to 3 so that MD engines can
check if PLUMED support this keyword before using it.

This fixes #100