Improved runtime loading procedure.
I made a few significant changes:

1. plumed_kernel_register is loaded with dlsym.

In this manner libplumedKernel does not depend anymore on
libplumed. This allows to load libplumedKernel library also
when the symbols defined in Plumed.c are not visible.

2. Introduced plumed_symbol_table struct.

This struct is included in the libplumedKernel library and
can be found from outside with dlopen. This struct contains
a version int (to allow for further extensions) and pointers
to the exported functions.

3. The loader (Plumed.c) has been significantly improved.

The new dynamic loader is expected to work also when
its symbols are not visibile since the
the main executable is not linked with -rdynamic.
If first tries to load libplumedKernel and, if this does
not work, libplumed (expected for old libplumedKernel not
independent of libplumed).
Then, it looks for the correct symbols, bypassing the automatic
registration procedure we used so far. Symbols are searched in
the plumed_symbol_table or, for backward compatibility,
by directly searching the plumed_plumedmain_cmd functions.

All the changes are backward compatible, in the sense that all
combinations (new/old loader with new/old libraries) are expected
to work. However, only the new loader works when its symbols are
not exported.