Fixed FIT_TO_TEMPLATE with COM/CENTER/GHOST
ActionWithVirtualAtoms should clean up the force on the virtual
atom when applied. Otherwise, actions computing the force on the
center of mass (such as FIT_TO_TEMPLATE) could compute it
wrongly when virtual atoms are defined.

I also modified a regtest so that it checks what happen when
using COM after FIT_TO_TEMPLATE