Simplified virial in some colvars
For DISTANCE, ANGLE, TORSION and PUCKERING,
I implemented something similar to GYR, where atoms coordinates
are first made whole, then the virial is computed assuming them whole.

Notice that modified coordinates are the local ones, so that the
change does not impact other actions.

Moreover, this required fixing numerical errors in a couple of regtests