Added autoconf program transform name
We usually allowed users to customize the name of the plumed
library using env var PLUMED_LIBSUFFIX. I changed it to the
standard autoconf behavior, that is using
--program-suffix
--program-prefix
at configure time.

PLUMED_LIBSUFFIX still works but is now deprecated and dumps a warning.
I will convert it to an error in a later version (2.3).