Histogram now uses new analysis architecture
I have changed the underlying base clases for analysis as the old
Analysis class was a mess.  This is now much tidier and more fit
for new purposes.  Nothing in the syntax was really changed - the change to
analysis/rt0/plumed.dat is simply because you now have to specify the
temperature if you calculate a histogram.