Rewrote analysis class
The old analysis class was reaching a stage where it was not fit
for purpose.  This is thus a total redesign in order to allow us
to write better algorithms for dimensionality reduction that are
flexible and easy for users to use.  I have removed some functionality
with this commit but I will re add it in subsequent commits
Also added information on analysis classes to developer documentation