Changed the way data is passed to reference configurations in analysis
The data is now passed via a pdb object.  This means we can now use
the same infractructure for setting up the ReferenceConfiguration as
we use in setting up the ReferenceConfigurations in paths and so on.
Projections meanwhile are attached to the ReferenceConfigurations
when they are required for printing.  So now a reference configuration
can be the pair of objects - i.e. the projections and the high dimensional
configuration.