Added new functionality for output pdb in analysis
You can now use COLLECT_FRAMES to collect pdb data from the trajectory
and output a pdb file that contains the underlying atomic coordinates
in pdb format even if you are not using the rmsd distance to measure
the dissimilarity matrix for dimensionality reduction.  Also fixed
a number of bugs.