Added plumed config tool
I implemented a shell-only tool "plumed config"

This tool can be used to probe the src/config/config.txt
configuration file.

So far it can be used e.g. as
plumed config has mpi
to enquire for an enabled feature.

It has some advantage wrt the --has-xxx options:

1. It is based on file src/config/config.txt,
which parses configure.ac and header files to find possible -D
and Makefile.conf to find activated -D.
This means that it is not necessary to manually
edit the src/config/Config.h and src/core/CLToolsMain.cpp
file to add new defines

2. it is shell only, thus works in cross compiled environment