Fix in enable-modules
Sometime there was problem with inconsistent on/off files. Now:

whenever all/none/reset is found, all on/off files are removed.
whenever one module is enabled, on file is added and off file is removed
the opposite when it is disabled

this should enforce consistency of on/off files.