Made symmetry functions that invove three atoms use less memory
In multicolvar when you have a triple loop it becomes very
expensive to store all the tasks in the triple loop in memory
(so that we can run over a single loop).  For actions where this
is very necessary I have thus made it so that you can run over a
double loop instead.  Inside each task of the double loop you then
run over the set of tasks that would have been in the double loop.