Allow bias to have derivatives
This implementation is not very satisfactory since it requires
to modify the implementation of each Bias to allow the bias component
to be itself biased.

Before merging it should be reimplemented in such a way that only
the component named bias has derivatives and it does it automatically from forces.