Possibility to choose atoms by name
I have implemented a minimalistic change, using the Molinfo infrastructure.
Namely, when one uses a word that is not a supported group names
(such as alpha,beta, etc) than plumed looks for a single atom with that name.

E.g.
phi-1
is the phi angle of residue one
CA-1
is the C alpha of residue one