Some small changes that allow for opitimisations of multicolvars
It is now possible to specify a different link cell cutoff for the
third bodies.  In addition, the way that connectivity is determined
in adjacency matrix actions works better given that we can now
operate with non-contact matrix contact matrices.  Also added some
cutoff stuff to HistogramBead.  This is currently not used in many
places