Values, made clearDerivatives faster
@all
clearDerivatives is a plumed bottleneck if one wants to work with many many
components (and probably also many collective variables). In a case with
a CV with 1000 components the code spent > 90% in clearDerivatives,
with this change it gets a little bit better but it is still very slow.
Do we need it?