Fixed parse of cmd string
Before this fix, using cmd() with an erroneous string was triggering
no error. Now this is fixed.

Notice that this is an old bug (introduced in 0a46f15d).
It should be harmless when using MD codes with proper interfaces,
but could be very dangerous if there is a typo in a cmd string.