Temporarily changed failing regtest
This regtest is failing on travis from time to time, in an
irreproducible way. I took note of the problem at issue #183

I temporarily change this regtest by:
1. moving COLVAR.reference to COLVAR.reference.dontcheck
2. adding HILLS.reference so that at least one file is checked.

Notice that this is not a solution of the problem, just a way
to avoid random failures on travis.

As soon as someone is willing to fix this, just put COLVAR.reference
back in place and push it to travis