Fixed some bugs that appear in analysis + reference
If you had an action that inherited from both ReferenceArguments and
ReferenceAtoms then there was a variable that appeared with the same name
in both so you can get conflicts.  In addition, RMSD object in optimal RMSD
needs clearing prior to setting.  Lastly made it so that the metric is printed
out in emdbedding files.