Fix in stopwatch
There was an assert checking that a stopwatch is not
started more then once. I replaced it with a counter
that, if a stopwatch is started N times, ignores the first
N-1 pause signals. This makes it possible to used Plumed.cmd
recursively. In practical terms, this makes it possible to use
DEBUG logRequestedAtoms
that was previosuly crashing the code.