Made plumed tolerate noEOL only in input
No EOL at the end of the file is only tolerated in input file
(e.g. plumed.dat) and not in other files (including HILLS files).

Indeed, when using multiple walkers, it is possible that transiently
the HILLS file from other walkers have incomplete lines.
With this fix, those lines are ignored and read at the next iteration.

This fixes #190