Fixed autoconf for gsl
I made it consistent with other libraries.

Now, plumed looks first for the required functions and, if not
found, adds a standard -l flag. This allows as in the other cases
to override the paths using the LIBS variable.

As a side effect, it might be that -lgslcblas does not get linked.
This could happen if linked blas already define a cblas interface
(as far as I know, this happens with mac for me and not with linux).