Fix problems with single precision blas/lapack on mac
On Mac, system blas/lapack returning single precision numbers
have an interface that is incompatible with the one included
in plumed. Notice that this does not happen when installing
plumed with macports, which uses vecLibFort to replace the
calls. Also notice that currently plumed only uses
double precision blas/lapack. Anyway, better to fix it for
the future.

In this commit I add:

- a regtest that fails when the bug is present
- a fix
- a test in configure that finds when to enable the fix

Notice that a macro gets defined when the fix is required for blas:
-D__PLUMED_BLAS_RETURNS_FLOAT=double
and an independent one for lapack
-D__PLUMED_LAPACK_RETURNS_FLOAT=double

This should allow to safely use internal lapack with external blas.