Doxygen 1.8.13
Notice that there are a few errors in some formulas.
To see all the warnings, look in the raw log from travis ci.
I also opened an issue right now (#213).

Anyway, I checked and the same errors were in the previous manual
(with doxygen 1.8.12). I thus upgrade to latest doxygen.

[makedoc]