Added framework to allow for a fix
This is a nasty workaround. Since FIT_TO_TEMPLATE needs to
access forces stored in Atoms, it is necessary that they are kept up
to date. With this fix, there is an extra function
allowToAccessGlobalForces()
that switch off an optimization that would otherwise make the
forces wrong when running in parallel