Move semantics for PLMD::Plumed
Notice that even though we assume c++11 to be available
in PLUMED 2.4 I here protected the definitions so that
they are only available with a c++11-supporting compilers
(see comments in the code).

This should not be necessary in other parts of the code,
since autoconf is already checking that c++11 is supported.