class ForwardDecl to change forward references to unique_ptr
I changed all the forward references to unique_ptr
using an auxiliary class (ForwardDecl).

Notice that in this manner the order of destructors
is exactly the reverse of the order of constructors.
This forces me to move a variable up in the layout
of PlumedMain to avoid a segfault.

This change is necessary to make plumed exception-safe.