Improved macports file
List of modules used in the Portfile is now generated automatically
to better deal with user contributed modules. Variants are called
mod_xxx
where xxx is the module name.

In addition, a variant +allmodules has been added to enable all modules.
This variant is used in the travis file to make sure all modules are tested.

I am not sure it is worthwhile to also automatize other variants, so I think
this closes #193