Implemented RECT keyword for METAD
Allows specifying multiple BIASFACTORs in a single file.

In addition, it triggers a citation to "altruistic" paper
by Spiwok et al when using METAD with RECT and WALKERS_MPI.
Notice that in this case the free-energy estimator is shared
rather than the bias itself, so as to allow "altruistic" updates
when walkers have different biasfactors (as in RECT).