Added autoconf flag --disable-static-patch
This flag disables the autoconf tests related to static patching.
The rationale is to be 100% sure that a wrong c++ library could not
be accidentally linked.

Notice however that this does not disable the "plumed patch --static"
flag. I think this should be fixed for the procedure to be more robust.

In the future, we could make this flag disabled by
default (to discourage static patching). Meanwhile, this is an option.