EDS: Fixed sign error in EDS implementation introduced in commit e2899f5
The means/centers substraction was reversed, inadvertantly introducing
a sign error. This affected adaptive runs with both covariance matrices and
without. Updated regtests and added additional information
in restarts used to debug this (means/std). Will help reproducibility when
restarting too. Accumulation was also too high for certain biases. It
was made to be 0, instead of 1.