Added support for PYTHON_BIN configure variable.
Use ./configure PYTHON_BIN=python2.7 to specify a version. Can
be either relative path or full path. If undefined, configure will
search from a list of names (currently, only name "python" is searched).

Notice that cython command is not searched, since it is not actually used
to build the interface. The test is rather done on the availability
of the Cython module and other required modules.

Moreover, I moved the test down in the list, after all the libraries