Passing variables to python build
Variables are passed as environment variables from Makefile,
where they can be constructed using Makefile variables.

This solves some problems I had with python2.7 not able to parse
correctly the Makefile.conf file. In addition, it looks like correct
directories are used also when plumed is already installed.

Also notice that the plumed executable is now only invoked
to retrieve the version number and that it is invoked with the
correct path, so that a previously installed plumed will
not be in the way.