Catch errors in PlumedMain::cmd()
I added a try/catch block so that any error that arises within a
cmd() call is also reported on the log file (if open).

In principle, this should allow removing all the duplicate code where we
both write on the log and on the plumed_merror.

Notice that in order to satisfy astyle the whole content of PlumedMain::cmd()
has been indented so appears as changed in this commit. Actually, it is only
indented.

This commit closes #319