Solved small bug in stored PLUMED_ROOT
The value of PLUMED_ROOT stored in statically patched MD code was
incorrectly set to the compilation directory rather than the installed one.

The only way to trigger the bug from regtest was by observing the behavior
of rt-make-exceptions when plumed was installed by macports (since in this
case the compilation directory is removed).