Separate repo for coverage scan
Coverage scan is now stored in a separate repository. This allows
generating the manual and the coverage scan in two different travis jobs.

Since this introduces a new travis job (for coverage scan), I removed another
job that was redundant (build with mpi and -O3; notice that we always use -O3
since v2.4). System lapack are now tested on the first job.

Fixes #348