unique_ptr: reference + mapping
I removed the remaining delete statements.

Notice that in order to make this change I had to change a bit the interface
of a few functions. The difficulty here is that there are parts of the
code storing a reference to a std::vector of pointers and those are affected
when the vector is transformed into a vector of unique_ptr.