Fixed program/lib names in install
In src/lib/install we create temporary versions of the to-be-installed
executable and libraries. I just changed them so that those temporary
versions are stored with their final $program_name name. This will
make it easier to link libplumed to libplumedKernel on linux as well
(as it was done for MacOS in b05560df36495f0afa87c967bb26cf9eaa1a2640).