Allow using absolute soname on Linux as well
On OSX, it is standard to hardcode the absolute path of a library in its install_name.
On Linux, the standard behavior so to only hardcode the soname and let
the system find the library. This actually requires setting LD_LIBRARY_PATH.

I added the possibility to use ./configure --enable-absolute-soname
to make Linux behave such as OSX.

I took the occasion to simplify considerably the linking of shared libraries.
Now the same command is used on Linux and OSX.