Allow remapping exceptions.
When using the C++ interface, this change remaps exception within plumed
to a class that inherits from std::runtime_error.
This allows them to be catched also from an MD code linked to
a different C++ library.

I also removed some cpp options, I think it is better to enforce
exceptions to be managed in this way.