Made python wrapper self-contained
It does not link to library anymore, but just includes Plumed.h file.
It thus could be compiled with different compilers and flags wrt plumed.

Notice that as of PLUMED 2.5 Plumed.h also contains the full implementation
of the dynamic loader, so it is sufficient to define the proper cpp
flags in order to recompile the loader and avoiding the need to link
plumedWrapper.a.

This basically allows the python wrapper to be selfcontained and installable
before having installed plumed.