Added py-plumed in MacPorts
I made a few more changes to setup.py in order to work with MacPorts.
This allowed me to add a temporary Portfile for py-plumed
that is now used in travis tests.

The plan is to submit this Portfile to MacPorts after we release v2.5

Notice that one test is duplicated (python/test) in order to be available
when using nosetests from python.