Fixed VIRIAL units and changed to clipped grad
The virial was not in the correct units and did not account for box size correctly. This was fixed, although non-orthorombic boxes are not treated exactly (see note in source). A warning is given to the user.

The gradients used a complex clipping procedure that attempted to clip the added step, which then unwound the addition to the accum.  However, the unwind wouldn't remove very large steps even if they  were clipped. The steps were then added to the accumulation, preventing further progress due to its large magnitude. This was  fixed and required to get virial results correct. This bug seemed to manifest as hitting large grads and then being stuck in a local minimum. This  didn't really matter in any but the largest gradient systems and only arose from large gradients from virial minimization. To fix this, gradients are just clipped. This change also makes it so that a first step will no longer be non-zero if the gradient is zero, which was the behavior previously with the overly-complicated clip. All regtests had to be revised due to that change.