I have successfully implemented all the parameters that were in the old Gyration.cpp file.
The parameters that I calculated in the previous commit differed by a
square root which I later found out was because the authers of [1]
prefered to work with Collective Coordinates that have dimensions of
length.

I successfuly reproduced these parameters:
1- GTPC_#        : Gyration Tensor Principal moment #, 1-3 in decending
		   order.

2- RADIUS        : Radius of Gyration, i.e. square root of the trace of
		   the Gyration Tensor.

3- ASPHERICITY   : The asphericity.
4- ACYLINDRICITY : The acylindricity.
5- RGYR_#        : The principal radii of gyration around the principal
		   axes.

But my results for the "TRACE" of the Gyration matrix differs by a factor
of 10!

Since the trace of the gyration tensor is calculated before the Eigen
values and all other parameters listed above, I have become suspicious
of the trace int the test outputs. Also, the relation between the trace
and the radius of gyration is a very simple one!

[1] Gyration- and Inertia-Tensor-Based Collective Coordinates for
Metadynamics. Application on the Conformational Behavior of Polyalanine
Peptides and Trp-Cage Folding, Jirí Vymetal, and Jirí Vondrasek, Journal
of Physical Chemistry, 2011, dx.doi.org/10.1021/jp2065612