Changes to example input files for CS2Backbone and Metainferences
I made these changes based on the errors I got when I parsed the code.
Can you please double check these are OK Max though.

Notice I also changed ../data/ in one place for data/ as it would be a pain to
get the syntax checking to work if this directory is in multiple different places
with different names.