Fixed warning with recent clang++
The warning is visible with clang 9.0 -Wdeprecated as:

warning: definition of implicit copy constructor for 'Invalid' is deprecated because it has a user-declared destructor

Notice that the copy constructors were removed here: fdf1a882bc037e9c250b6caef5c660849d23ab4e, and this commmits
appears also in v2.5 and v2.6.

This fix should be probably backported to v2.5 and v2.6.