DRR: add a new option for specifying the output name of --merge. (#662)
The old behavior of generating output filename for --merge is not
suitable in the cases that the input filenames contain "/", which
indicates the input files are in different sub-directories. Simply
joining the input filenames as the name for output does not work,
so I add a new option --merge_output for these cases. After applying
this patch, the output name can be specified as following:

plumed drr_tool --merge dir1/win.drrstate,dir2/win.drrstate
--merge_output xxx