Attempting to add GHBFIX and regtest-ghbfix
Description

This is an attempt to add the colvar GHBFIX, which essentially is a switching function for atom pairs, being -1 for small distances and 0 for large distances with a smooth interpolation between them.

This variable computes the ghbfix interaction among GROUPA and GROUPB as defined in Kührová et al.(2016): Computer Folding of RNA Tetraloops: Identification of Key Force
Field Deficiencies. JCTC.
It can be scaled as desired by specifying interaction scaling parameters and energy units.