Decoding nested exceptions in the C wrappers
In the C wrappers exceptions are stored in the plumed_error
structure. This structure now contains a linked list.
Finalizing the outer error is sufficient to finalize the whole
list. Thus, there is no change from the user perspective, except
that it is now possible to scroll the list of nested exceptions.