Added -fno-gnu-unique
This option should solve once for all the problem with gnu unique symbols.
I am not 100% this option will not introduce troubles linking other C++
libraries, and thus I leave the possibility to remove it with
./configure --disable-no-gnu-unique

However, problems are very unlikely considering that this is the
same behavior of icpc and clang.

See #549