Started adding vector/matrix/grid stuff into Value from hack-the-tree
In hack-the-tree you can pass vectors/matrices and grids through Values.
This necessitates some changes to the Value class.  In particular inputForce
becomes a vector and similarly the double value becomes a vector called data.  I avoid
having a confusing vector called derivatives (which is not always the derivatives).
If I have a scalar I store the value and the derivatives in the data vector.  data
thus has a length of 1 + number of derivatives.  The first element is the value of the scalar.

Notice that I also have functionality to set the value to be a constant that is calculated once and
only once during the calculation