Added functionality for passing data to and from PLUMED from hack-the-tree
This functionality introduces two new actions into plumed:

PUT = an ActionWithValue that transfers some data from a pointer that is passed from the MD code to a PLMD::Value.
GET = an ActionWithArgument that transfers data from a PLMD::Value to a pointer than is passed from the MD code

Notice that you can also pass a pointer to the forces on a object that is passed to PUT.  The forces from value
are then added to the MD code's force pointer when you call the apply method.

To demonstrate how this functionality works I have replaced all the stuff for passing extraCVs to PLUMED with PUT
commands

I have amm also using GET in place of the old DataFetchingObject class, which has now been deleted