Changes to pass timestep from MD code to PLUMED through a PUT action
The timestep is now passed to PLUMED using a PUT action.  The timestep is
thus stored in a value called timestep that can be accessed using the label
timestep.  Notice that changes to the object that collects the data from the MD
code were required here because the timestep is passed directly (and not via a
pointer).  To resolve this there is a variable in DataPassingObject that can hold
a double that is passed from the MD code.