Passing cell vectors through PUT action
This commit modifies the code so that cell vectors are passed through a
PUT action.  The virial is then passed as the forces that any forces on the
input value are added on.  To deal with the fact that the virial and cell should only
be passed from one of the MD codes processes there is a special new class PbcAction
that deals with passing the cell vectors.  Methods from this class are also invoked
when the box parameters are sought elswhere in the code.

Notice that I have also added the DomainDecomposition class from hack-the-tree here.
Much of the functionality in this class is not yet used.  It also needs to be modified
before it can be used here.  I felt as we merge hack-the-tree it is easier to start out
with a code that resembles the one in hack-the-tree.