Passing kBT through a PUT action
kBT is now passed from the MD code into a PLMD::Value by passing the value
to a PUT action.  Whenever an action needs kBT you can call a method in action
called getkBT.  This reads in the keyword TEMP (if it is registered).  If TEMP
is present the value read for this flag is multiplied by Boltzmann's constant.  If
TEMP is not present then the kBT value that was passed from the MD code to PLUMED is
returned.