Fixed mechanism for passing energy between MD code and PLUMED
The energy is now passed through a specialised PUT action.  When apply is called
for this PUT action the rescaleForces method of the position and forces PUT actions
are called.  This allows one to adjust the forces on the atoms in the correct way due
to a force that acts on the value of the energy