Moved duplicated functionality for getting forces from Values in Function and Colvar to ActionWithValue
The way that you get the forces from the Values is the same in Function and in Colvar.  I thus created a
method called getForcesFromValues in ActionWithValue and used this method in Function and Colvar.  I also
rewrote the interface for the methods setForcesOnAtoms and addForcesToArguments.  You now passed a reference
to an unsigned which keeps track of where you are in the forcesForApply array that is passed to this functions.