Virtual atom positions are now passed in three PLMD::Value
ActionWithVirtualAtom now creates 5 values: the x, y and z position of the atom
and the mass and charge of the atom.  This replaces the mechanism whereby additional
atoms were created in the Atoms class.  ActionAtomistic creates vectors containing
pointers to the PLMD::Values that are created by the PUT actions that pass positions from
the MD code and the Virtual Atom actions.  If virtual atoms are needed a dependency on the
virtual atoms action is created in requestAtoms.