Wrote code to ensure Values that should be calclated during update are calculated during update
In hack the tree all averages are stored in PLMD::Value objects.  Furthermore, you can
calculate a function of an average that is stored in a PLMD::Value.  This function should
not be calculated when you call calculate, however.  It should be called during the update
step.  This little bit of code allows plumed to deal with functions of averages