Rewrote functions so that we can use them with vector, grid and matrix as input eventually.
It is useful to be able to use the CUSTOM, SORT and Piecewise functions with inputs other than scalars.
If a multiple vectors are passed to CUSTOM, for example, you can compute the defined function element wise.
This mechanism is used extensively when implementing complicated CVs in the hack-the-tree branch.  In order
for this to appear simple to users a shortcut action is called when you create one of the functions above.
This shortcut action decides whether the input is a scalar, vector, matrix or grid and calls the version of the
action that is appropriate to the input arguments.  FunctionOfScalar, FunctionOfVector, FunctionOfMatrix and
FunctionOfGrid template classes are used to create CUSTOM_SCALAR, CUSTOM_VECTOR, CUSTOM_MATRIX and CUSTOM_GRID
actions.  The type of function being used is the template parameter and must inherit from the class FunctionTemplateBase.

In this commit I adjusted the code so that this new mechanism is used for scalar function.  We cannot compute vectors yet
with PLUMED so there is no point in adding the FunctionOfVector code from hack-the-tree yet.  However, I will add this in a
future commit that will take place shortly.

Changes to regtests are simply because there is an additional action in the input.  Action with labels such as @3 thus change
to having labels like @4