Added shortcut action to create basic Colvars as I will use them in MultiColvar to create vectors of these basic quantities
ANGLE, DIPOLE, DISTANCE, POSITION and TORSION currently calculate 1 or more scalar quantities.
It would be easy to imagine another version of these colvars that takes multiple groups of atoms
and that calculates multiple angles or multiple distances.  This is what MultiColvar does.  In hack-the-tree,
however, it is implemented as a wrapper round the underlying Colvars.  Eventually you will be able to implement
a Multicolvar and a Colvar just by implementing a Colvar