Started reimplimenting MultiColvar in a much simpler way
A MultiColvar in hack-the-tree is a vector of colvars of a particular type.  You specify
what is to be calculated for a distance multicolvar by doing:

DISTANCE ATOMS1=1,2 ATOMS2=3,4 ATOMS3=5,6

A vector containing three distances is then computed.  I have implemented this by writing a
general templace class called MultiColvarTemplate.  The template parameter of this class is one
of the regular Colvars that we have implementations.  You can thus reuse the code for Distance
to calculate the vector with three distances above.

To make this work I have to move some of the functionality in the underlying colvars to three
static functions. To be clear, you can still implement Colvars in the old way.  Hopefully
the way to implement a Colvar so that it is automatically a MultiColvar is relatively self explanatory.