Added code to calculate functions of Vectors
Now that we can calculate a vector of CVs using MultiColvarTemplate we should also
be able to calculate a function of the vectors that the MultiColvarTemplate function
outputs.  The same set of functions (and more) that are available for scalars can be used
to calculate functions of vectors.  We thus defined a FunctionOfVector class here.

I also modified the rt-basic-print regtest to test the FunctionOfVector class