Added command land tool that provides a mermaid graph that illustrates how data passes through the code
With PLUMED passing Vector, matrices, grids and scalars between actions it is useful to have some
illustration of how data is being passed about inside the code.  This commit provides a tool that
can be used to visualise inputs.  Show_graph draws a graph.  Each node in the graph is an action
and the connections between nodes illustrate the PLMD::Value objects that are passed between Actions.
If a connection is coloured black then a scalar is passed.  If it is blue, red or green then a vector,
matrix and grid respectively is being passed.  Lastly violet connections are used to show how atoms are
passed between actions.