Added functionality from hack-the-tree for constact matrices (and coordination numbers)
A useful first step in calculating many CVs in condensed matter is to compute a matrix
in which element ij tells you if there is a "bond" between atoms/molecule i and atom/molecule
j.  If this matrix is multiplied by a vector of ones you get the coordination numbers of the
molecules.  You can calculate SPRINT coordinates from the eigendecmposition of this matrix.
You can also find clusters of atoms by doing a clustering on the adjacency graph that this matrix
describes.

I have written a base class here to calculate these adjacency matrices.  This base class can also
be used to do products of matrices and vectors (to get coordination numbers).