Changes to make clearInputForces faster when you pass small number of atoms
In PLMD::Value clearInputForce does a std::fill over the whole of the inputForce
vector to clear the forces that are acting on a vector.  This is a problem when you
have forces on the atomic positions because inputForce is a std::vector that has natoms
component. In many cases a lot of those forces are zero because only a small subset of atom
positions are used by PLUMED.  To get around this I have added some methods to ensure that only
only the elements of inputForce that correspond to atom positions that have been passed by PLUMED
are set to zero.