Reimplemented functionality for calculating the number of particles in a particular volume and the average value of a quantity in a volume
We can introduce a new type of ActionVolume object that takes in a series of atom positions and returns a vector that is one if each of those
atoms are within a particular part of the box and that is zero otherwise.  If you take the sum of these vectors you get the number of atoms
that is in the particular part of the box of interest.  However, you can also multiply these ones and zeros by vectors of symmetry functions
and hence determine the average value of a particular symmetry function in a particular part of the box.  This functionality was implemented
in previous versions of PLUMED but now it is done in a much more transparent way.