First commit of module for molecular crystalline order from Jake McKibbin, Erik Santiso and me
This module will provide tools for comparing the relative orientations of molecules in a medium with the
orientations that they have in the ideal crystal structure of the solid.  Some of the functionality for
this sort of CV is already available in PLUMED.  Tests to ensure that you can use these features in the way
that Jake and Erik use them have thus been included in this commit.  This commit also includes the beginnings
of a class for calculating quaternions.  This class will be used when determining the orientations of the molecules.