Added quaternion product matrix action
This is used to calculate relative orientations of molecules in the ROPS order parameter from Trout and Santiso.
In Trout and Santiso's work they use a different method for calculating the quaternion than we used in our original implementation
of the quaternion.  We thus changed the way the quaternion is calculated so that what we do is consistent with what is done in those works.
In practise this means the signs of the imaginary components of the quaternions change.  This is why there are changes in the regression tests