Improved test with multiple threads
In the previous version, an error leading to an exception thrown in a thread
was reported as a crash.

With this fix, errors (per thread) are stored and reported in the output.
This makes the possible test failure easier to interpret.
In particular, the failure discussed in #990 in rt-make-threads
should be now visible as a numerical error.