Optimized dynamic_cast in core
For classes that are part of the core with can use a virtual function,
which is faster than a dynamic_cast
In this commit I got rid of most dynamic_cast's in the core of the code.
These virtual functions might be used profitably also in other parts of the code.

Notice that:
- this makes sense only for classes that are in the core of the code. Action class
  should not be aware of extensions, or we will loose in modularity.
- this might have impact on v2.9 as well (before htt). However, the impact on htt
  is expected to be larger since it uses actions to pass information from the MD code
  and because all these actions use dynamic_cast
- With this single commit, and a simple test with a 20 atoms system, 4 of which used in a biased
  CV, I can gain ~ 15% speed on my laptop.