Optimized getWords in PlumedMain
I basically reimplemented a simpler version of getWords which
has some limitation:
- only understand space (" ") as separator (no newlines or tabs)
- do not understand braces

This implies we cannot anymore do:
plumed_cmd(p,"{init}")
or
plumed_cmd(p,"init\t")
I think we never meant plumed_cmd to be used in this way. If we want
to pass more structured strings, we can always pass them as the second
argument. So I guess this is not a problem.

Another way to make it faster is to avoid allocations. Instead
of returning a std::vector<std::string>, we accept
a reference to a preallocated small_vector<std::string_view>.
This should require literally zero new allocations, as long as the number
of words is low (for us it's always 0, 1 or 2). This required some modifications
to the PlumedMain code, because
- these std::string_view should be converted to std::string in some cases (with a subsequent allocation)
- I had to modify the word_map object to be able to access it with a std::string_view

With this optimization, and a simple test with a 20 atoms system, 4 of which used in a biased
CV, I can gain ~ 20% speed on my laptop