Added benchmark command line tool
This tool runs plumed using a trajectory that is hard coded in the source code.
It essentially works very much like driver, except the trajectory is fixed.
You can use this to benchmark plumed calculations.  The advantage of having the
separate code is the the profiles you get from this tool are not influenced by lots of I/O.