Fix for gcc 8
Unfortunately it seems gcc 8 has a bug and cannot recognize
a noexcept statement.

My understanding is that this statement can be removed since
it would just lead to a worse performance in some specific operation
on some types of objects, but not to errors. I thus added a check
so that this statement is removed by the preprocessor when using
GCC 8 or less.