Improved behavior of plumed-runtime on Linux
I am making this change to allow the plumed-runtime command
to find libplumedKernel.so more easily. To do this,
I exploit the search path of dlopen, which is basically the same used
by the dynamic linker when linking libraries normally.

Before this fix, only the "non installed version" was functional,
thanks to the fact that PLUMED_KERNEL env var is set in sourceme.

In this commit:
- I removed the need for the PLUMED_KERNEL to be an exisiting file.
  This allows leveraging on the search path used by the operating system.

- If the user does not specify PLUMED_KERNEL, and if the path is not
  overridden at compilation time, I set it to lib(program_name)Kernel.so

These two changes together make a default search which is the proper
file name (libplumedKernel.so) in the LD_LIBRARY_PATH
(which is where the file should be installed).

In order to cover the case of user not setting LD_LIBRARY_PATH
but rather using an rpath, I made sure that the rpath is also added
to the plumed-runtime installed command. This should make the
plumed-runtime command functional "as is" also in conda installs
(to be tested)

On MacOS things work differently so I would have to address them
separately